+++ src/Credfeto.ChangeLog.Tests/ChangeLogUpdaterPreambleTests.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Credfeto.ChangeLog.Models;
 using Credfeto.ChangeLog.Services;
 using FunFair.Test.Common;
+using FunFair.Test.Infrastructure.Mocks;
 using Xunit;
 
 namespace Credfeto.ChangeLog.Tests;
@@ -95,7 +96,8 @@ Releases that have at least been deployed to staging, BUT NOT necessarily releas
                     await ParseAsync(ChangeLogWithoutPreamble),
                     "1.0.0",
                     pending: true,
-                    Language
+                    Language,
+                    MockDateTimeSources.Past
                 )
             )
         );